Disable glibc macros for character class functions.
First off, it's hard to understand glibc's macro expansion for these
functions. It turns, say, `isdigit(c)` into this:

    ((*__ctype_b_loc ())[(int) ((c))] & (unsigned short int) _ISdigit)

Getting `isdigit(c)` out instead is much more clear.

Second, glibc's enum that contains `_ISdigit` and so on is defined
differently based on the endianness of the target architecture, and on
little-endian targets, each definition involves a ternary conditional.
Corrode translates that happily but current Rust doesn't allow
if-expressions in constant expressions, and rejects the generated
definition.

So by forcing glibc to use regular function calls instead of macros, we
get (1) easier-to-read code that (2) actually works.