Simplify no-op casts and fix special casts.
If the C source casts an expression to the same type it already
evaluated to, we can just discard the cast. I didn't want to do that on
the grounds that the programmer might have been trying to document
something's type by casting it, but I've decided that the generated Rust
will be more clear with fewer casts.

Otherwise, it may be casting the expression to a type that it can't be
simply coerced to in Rust.

In both cases, the right logic is already in `castTo`, so just call
that.