Fix zero-initialization for static variables.
Static-lifetime variables that do not have an initializer in the C
source need an explicit zero-equivalent initializer when translated to
Rust.

I had previously kludged this so the constructed initializer was always
just the integer literal "0", because it seemed tricky to thread the
right data around to the right places. Turns out, it wasn't that tricky.